safer migration loading, some speedups
we no longer run each migration in noop mode to get at the triggers
(using an AR::Migration monkey-patch)... this was uber-dangerous since
we only intercepted things that fell through to the connection (e.g.
create_table or execute), and not other stuff (e.g. Foo.update_all or
MyWebService.call). we now use ruby_parser to extract triggers (or in
the case of schema.rb, just a simple regex)

added some little regex checks to model and migration loading so that
we don't bother loading or parsing classes that have no triggers