of(*columns) support, fixes #32
when creating an update trigger, you can use the new `of(*columns)` method
to indicate should only run if one of the listed columns is included in
the update statement. this is a more concise alternative to writing
`where("NEW.col <> OLD.col OR ...")`

in sqlite and newer versions of postgres, this is implemented as an actual
`OF ...` clause. other platforms will get an inferred/coalesced `WHEN`
clause or `IF` statement in the trigger body

note that it is not precisely equivalent to a `where("NEW.col <> ...")`
call, since `OF ...` clauses will run whether or not the values have
actually changed. the mere inclusion of the column in the statement is
sufficient to trigger it.

e.g. "UPDATE foos SET bar = bar" will trigger (ha) an `ON bar` clause, but
not an `IF NEW.bar <> OLD.bar` statement.