Add configuration class for custom adapters (#135)
The list of supported adapters is hardcoded, which makes it difficult to
use this gem on custom adapters.

We use `postgresql_proxy` from `active_record_proxy_adapters`, for
instance. Currently, we need to monkey patch the gem and hot swap the
constant `HairTrigger::POSTGRESQL_ADAPTERS` when the application boots
to make it work with the proxy adapter.

Although it works, it's not ideal.

I've added a new configuration class whose values will default to the
current ones (`%i[mysql2 trilogy]` for `mysql_adapters`, `%i[postgresql
postgis]` for `postgresql_adapters` , and `%i[sqlite litedb]` for
`sqlite_adapters`), but allow the user to add custom adapters for each.

I don't expect any breaking changes here as the values fall back to the
existing constants, but please double check. Thanks!